updating review student answers report

+++ BLL/Entities/StudentExam.cs
@@ -4,7 +4,7 @@
     {
         public Student student { get; set; }
         public Exam exam { get; set; }
-        public int? grade { get; set; }
+        public decimal? grade { get; set; }
         public char? status { get; set; }
         public int courseid {  get; set; }
